refactor: optimisation around stoptimes
- moved stoptime related functionality into new core:data:stoptime module
  - will feature all the different realtime stoptime sources to be
      integrated later
- create proper database schema for future migrations
- deduplicate trips into stoppingpatterns, since many trips share the
  exact same stopping pattern
  - stoptimes are now linked to stoppingpatterns instead
  - stoppingpattern ids are generated from a hash composed of all stoptimes
- stoptimes now use deltas for arrival time to save space

+++ server/gtfs/src/main/kotlin/moe/lava/banksia/server/gtfs/structures/GtfsStopTime.kt
@@ -10,7 +10,7 @@ internal data class GtfsStopTime(
     val arrival_time: String,
     val departure_time: String,
     val stop_id: String,
-    val stop_sequence: Int,
+    val stop_sequence: Long,
     val stop_headsign: String,
     val pickup_type: Int,
     val drop_off_type: Int,
